Configurando AreaEntityMap
