+++ Mda/Mda.Domain/Entities/Area.cs
@@ -14,6 +14,7 @@ namespace Mda.Domain.Entities
         public int NotaInicio { get; set; }
         public int NotaFim { get; set; }
         public int Ano { get; set; }
+        
 
     }
 }
